+++ package.json
@@ -44,7 +44,7 @@
     "formidable": "^3.5.1",
     "graphql-request": "^6.1.0",
     "lucide-react": "^0.316.0",
-    "maci-cli": "0.0.0-ci.d9ad7f8",
+    "maci-cli": "0.0.0-ci.753ded3",
     "next": "^14.1.0",
     "next-auth": "^4.24.5",
     "next-themes": "^0.2.1",
