+++ src/contexts/types.ts
@@ -1,0 +1,30 @@
+import { type ReactNode } from "react";
+import { type TallyData, type IGetPollData } from "maci-cli/sdk";
+
+export interface IVoteArgs {
+  voteOptionIndex: bigint;
+  newVoteWeight: bigint;
+}
+
+export interface MaciContextType {
+  isLoading: boolean;
+  isEligibleToVote: boolean;
+  initialVoiceCredits: number;
+  votingEndsAt: Date;
+  stateIndex?: string;
+  isRegistered?: boolean;
+  pollId?: string;
+  error?: string;
+  pollData?: IGetPollData;
+  tallyData?: TallyData;
+  onSignup: (onError: () => void) => Promise<void>;
+  onVote: (
+    args: IVoteArgs[],
+    onError: () => Promise<void>,
+    onSuccess: () => Promise<void>,
+  ) => Promise<void>;
+}
+
+export interface MaciProviderProps {
+  children: ReactNode;
+}
