+++ src/features/projects/components/ProjectsResults.tsx
@@ -4,10 +4,12 @@ import Link from "next/link";
 import { InfiniteLoading } from "~/components/InfiniteLoading";
 import { getAppState } from "~/utils/state";
 import { useResults, useProjectsResults } from "~/hooks/useResults";
+import { useMaci } from "~/contexts/Maci";
 import { ProjectItem, ProjectItemAwarded } from "./ProjectItem";
 
 export function ProjectsResults() {
-  const projects = useProjectsResults();
+  const { pollData } = useMaci();
+  const projects = useProjectsResults(pollData);
   const results = useResults();
   const appState = getAppState();
